+++ uniffi_bindgen/src/bindings/python/templates/ObjectTemplate.py
@@ -3,28 +3,36 @@ class {{ obj.name()|class_name_py }}(object):
     {%- when Some with (cons) %}
     def __init__(self, {% call py::arg_list_decl(cons) -%}):
         {%- call py::coerce_args_extra_indent(cons) %}
-        self._handle = {% call py::to_ffi_call(cons) %}\
+        self._pointer = {% call py::to_ffi_call(cons) %}
     {%- when None %}
     {%- endmatch %}
 
     def __del__(self):
-        rust_call_with_error(
-            InternalError,
-            _UniFFILib.{{ obj.ffi_object_free().name() }},
-            self._handle
-        )
+        # In case of partial initialization of instances.
+        pointer = getattr(self, "_pointer", None)
+        if pointer is not None:
+            rust_call_with_error(
+                InternalError,
+                _UniFFILib.{{ obj.ffi_object_free().name() }},
+                pointer
+            )
+
+    # Used by alternative constructors or any methods which return this type.
+    @classmethod
+    def _make_instance_(cls, pointer):
+        # Lightly yucky way to bypass the usual __init__ logic
+        # and just create a new instance with the required pointer.
+        inst = cls.__new__(cls)
+        inst._pointer = pointer
+        return inst
 
     {% for cons in obj.alternate_constructors() -%}
     @classmethod
     def {{ cons.name()|fn_name_py }}(cls, {% call py::arg_list_decl(cons) %}):
         {%- call py::coerce_args_extra_indent(cons) %}
         # Call the (fallible) function before creating any half-baked object instances.
-        handle = {% call py::to_ffi_call(cons) %}
-        # Lightly yucky way to bypass the usual __init__ logic
-        # and just create a new instance with the required handle.
-        inst = cls.__new__(cls)
-        inst._handle = handle
-        return inst
+        pointer = {% call py::to_ffi_call(cons) %}
+        return cls._make_instance_(pointer)
     {% endfor %}
 
     {% for meth in obj.methods() -%}
@@ -33,12 +41,12 @@ class {{ obj.name()|class_name_py }}(object):
     {%- when Some with (return_type) -%}
     def {{ meth.name()|fn_name_py }}(self, {% call py::arg_list_decl(meth) %}):
         {%- call py::coerce_args_extra_indent(meth) %}
-        _retval = {% call py::to_ffi_call_with_prefix("self._handle", meth) %}
+        _retval = {% call py::to_ffi_call_with_prefix("self._pointer", meth) %}
         return {{ "_retval"|lift_py(return_type) }}
 
     {%- when None -%}
     def {{ meth.name()|fn_name_py }}(self, {% call py::arg_list_decl(meth) %}):
         {%- call py::coerce_args_extra_indent(meth) %}
-        {% call py::to_ffi_call_with_prefix("self._handle", meth) %}
+        {% call py::to_ffi_call_with_prefix("self._pointer", meth) %}
     {% endmatch %}
     {% endfor %}