+++ uniffi_bindgen/src/bindings/ruby/templates/RustBufferBuilder.rb
@@ -105,6 +105,14 @@ class RustBufferBuilder
     write v
   end
 
+  {% when Type::Object with (object_name) -%}
+  # The Object type {{ object_name }}.
+
+  def write_{{ canonical_type_name }}(obj)
+    pointer = {{ object_name|class_name_rb}}._uniffi_lower obj
+    pack_into(8, 'Q>', pointer.address)
+  end
+
   {% when Type::Enum with (enum_name) -%}
   {%- let e = ci.get_enum_definition(enum_name).unwrap() -%}
   # The Enum type {{ enum_name }}.