+++ uniffi_bindgen/src/bindings/swift/templates/EnumTemplate.swift
@@ -1,7 +1,7 @@
 
 // Note that we don't yet support `indirect` for enums.
 // See https://github.com/mozilla/uniffi-rs/issues/396 for further discussion.
-public enum {{ e.name()|class_name_swift }}: Equatable, Hashable {
+public enum {{ e.name()|class_name_swift }} {
     {% for variant in e.variants() %}
     case {{ variant.name()|enum_variant_swift }}{% if variant.fields().len() > 0 %}({% call swift::field_list_decl(variant) %}){% endif -%}
     {% endfor %}
@@ -39,3 +39,7 @@ extension {{ e.name()|class_name_swift }}: ViaFfiUsingByteBuffer, ViaFfi {
         }
     }
 }
+
+{% if ! e.contains_object_references(ci) %}
+extension {{ e.name()|class_name_swift }}: Equatable, Hashable {}
+{% endif %}