+++ uniffi_bindgen/src/scaffolding/templates/ObjectTemplate.rs
@@ -1,13 +1,14 @@
-// For each Object definition, we assume the caller has provided an appropriately-shaped `struct`
-// with an `impl` for each method on the object. We create a `ConcurrentHandleMap` for safely handing
-// out references to these structs to foreign language code, and we provide a `pub extern "C"` function
+// For each Object definition, we assume the caller has provided an appropriately-shaped `struct T`
+// with an `impl` for each method on the object. We create an `Arc<T>` for "safely" handing out
+// references to these structs to foreign language code, and we provide a `pub extern "C"` function
 // corresponding to each method.
+//
+// (Note that "safely" is in "scare quotes" - that's because we use functions on an `Arc` that
+// that are inherently unsafe, but the code we generate is safe in practice.)
 //
 // If the caller's implementation of the struct does not match with the methods or types specified
 // in the UDL, then the rust compiler will complain with a (hopefully at least somewhat helpful!)
 // error message when processing this generated code.
-{% let handle_map = format!("UNIFFI_HANDLE_MAP_{}", obj.name().to_uppercase()) -%}
-
 
 {% if obj.uses_deprecated_threadsafe_attribute() %}
 // We want to mark this as `deprecated` - long story short, the only way to
@@ -23,25 +24,27 @@
 fn uniffi_note_threadsafe_deprecation_{{ obj.name() }}() {}
 {% endif %}
 
-uniffi::deps::lazy_static::lazy_static! {
-    #[doc(hidden)]
-    static ref {{ handle_map }}: uniffi::ffi::handle_maps::ArcHandleMap<{{ obj.name() }}>
-        = Default::default();
-}
-
-    {% let ffi_free = obj.ffi_object_free() -%}
-    #[doc(hidden)]
-    #[no_mangle]
-    pub extern "C" fn {{ ffi_free.name() }}(handle: u64) {
-        let _ = {{ handle_map }}.delete_u64(handle);
+{% let ffi_free = obj.ffi_object_free() -%}
+#[doc(hidden)]
+#[no_mangle]
+pub extern "C" fn {{ ffi_free.name() }}(ptr: *const std::os::raw::c_void) {
+    // We mustn't panic across the FFI, but also can't report it anywhere.
+    // The best we can do it catch, warn and ignore.
+    if let Err(e) = std::panic::catch_unwind(|| {
+        assert!(!ptr.is_null());
+        {#- turn it into an Arc and explicitly drop it. #}
+        drop(unsafe { std::sync::Arc::from_raw(ptr as *const {{ obj.name() }}) })
+    }) {
+        uniffi::deps::log::error!("{{ ffi_free.name() }} panicked: {:?}", e);
     }
+}
 
 {%- for cons in obj.constructors() %}
     #[allow(clippy::all)]
     #[doc(hidden)]
     #[no_mangle]
     pub extern "C" fn {{ cons.ffi_func().name() }}(
-        {%- call rs::arg_list_ffi_decl(cons.ffi_func()) %}) -> u64 {
+        {%- call rs::arg_list_ffi_decl(cons.ffi_func()) %}) -> *const std::os::raw::c_void /* *const {{ obj.name() }} */ {
         uniffi::deps::log::debug!("{{ cons.ffi_func().name() }}");
         {% if obj.uses_deprecated_threadsafe_attribute() %}
         uniffi_note_threadsafe_deprecation_{{ obj.name() }}();